Refactored LruCache
